Move extensions tests (#7318)
