+++ src/extensions/billingMigrationHelper.spec.ts
@@ -1,11 +1,11 @@
 import { expect } from "chai";
 import * as sinon from "sinon";
 
-import { FirebaseError } from "../../error";
-import * as nodejsMigrationHelper from "../../extensions/billingMigrationHelper";
-import * as prompt from "../../prompt";
-import { ExtensionSpec } from "../../extensions/types";
-import { cloneDeep } from "../../utils";
+import { FirebaseError } from "../error";
+import * as nodejsMigrationHelper from "./billingMigrationHelper";
+import * as prompt from "../prompt";
+import { ExtensionSpec } from "./types";
+import { cloneDeep } from "../utils";
 
 const NO_RUNTIME_SPEC: ExtensionSpec = {
   name: "test",
